Added a second attack type to replace old self-damage shift function
