+++ Assets/Creatures/Exposion.psd
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:a7f661ab66fd3c18975aa50c2a0843ed8fd80baeace968a36cbe0b54fea2ec08
+size 69274
