+++ Assets/Scripts/ArenaHandler.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
-using Attacks;
+using AttackTypes;
 
 public class ArenaHandler : MonoBehaviour
 {
@@ -33,6 +33,7 @@ public class ArenaHandler : MonoBehaviour
         var playerCreature = playerCreatureGO.GetComponent<Creature>();
         SetupCreature(playerCreature, "Cubey", 10, 69, HumanPlayer);
         playerCreature.LearnAttack(new ConeAttack());
+        playerCreature.LearnAttack(new CenteredAttack());
 
         var enemyCreature = enemyCreatureGO.GetComponent<Creature>();
         SetupCreature(enemyCreature, "Cylindork", 5, 80, NPCPlayer);
