+++ Assets/Scripts/AttackList.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 1d7e1d3c724f4b3f9481380346a2e04f
+timeCreated: 1616995466
