+++ Assets/Scripts/AttackTypes/CenteredAttack.cs
@@ -1,0 +1,17 @@
+﻿using System.Collections;
+using System.Runtime.InteropServices.ComTypes;
+using UnityEngine;
+
+namespace AttackTypes
+{
+    public class CenteredAttack : IAttack
+    {
+        public IEnumerator Attack(Vector2 direction, Transform[] exitPoints, GameObject attackPrefab)
+        {
+            var centered = Object.Instantiate(attackPrefab, exitPoints[6].position, Quaternion.identity);
+            yield return new WaitForSeconds(.25f); 
+    
+            Object.Destroy(centered);
+        }
+    }
+}
